(hel-678) hel-678 decompose table into parts
