+++ src/frontend/ui/parametrage-utilisateurs/UsersListPage/Pagination/PaginationBtn/PaginationBtn.module.css
@@ -1,6 +1,7 @@
 .pagination_container {
   margin-top: 30px;
-
+  display: flex;
+  justify-content: center;
   user-select: none;
 }
