+++ src/frontend/ui/recherche-avancee/resultat-recherche-avancee/ResultatRecherchePlaceHolderText.tsx
@@ -1,6 +1,6 @@
+import { useDependencies } from "../../commun/contexts/useDependencies";
 import "@gouvfr/dsfr/dist/component/segmented/segmented.min.css";
-import { useDependencies } from "../commun/contexts/useDependencies";
-import styles from './RechercheAvanceeFormulaire.module.css'
+import styles from '../RechercheAvanceeFormulaire.module.css'
 
 export const ResultatRecherchePlaceholderText = () => {
     const { wording } = useDependencies();
